pyproject.toml
[project]
name = "prodrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google>=3.0.0",
    "google-genai-haystack>=3.11.0",
    "haystack-ai>=2.26.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.14",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langfuse>=3.14.5",
    "langfuse-haystack>=4.1.0",
    "langgraph>=1.1.3",
    "langsmith[claude-agent-sdk]>=0.7.22",
    "pinecone-haystack>=6.1.2",
    "pypdf>=6.9.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.44.1",
]
